Refactor file descriptor handling and path imports
This commit refactors the handling of file descriptors and updates the code to use absolute paths instead of relative paths. This has been done across the file_system, task, and graphics modules, removing redundancies and enhancing code readability and maintainability. The file descriptor's read method has been moved from the deleted 'file_descriptor.cpp' file to 'fat.cpp' for better organization of related functionalities.
