+++ kernel/file_system/CMakeLists.txt
@@ -1,6 +1,5 @@
 set(FILE_SYSTEM_SOURCE_FILES
         fat.cpp
-        file_descriptor.cpp
 )
 
 add_library(UchosFileSystem ${FILE_SYSTEM_SOURCE_FILES})
