+++ kernel/graphics/color.cpp
@@ -1,4 +1,5 @@
 #include "graphics/color.hpp"
+#include <cstdint>
 
 Color::Color(uint8_t r, uint8_t g, uint8_t b) : r_(r), g_(g), b_(b) {}
